Corrected and cleaned some imports, moved and tweaked PyPi versioning
Some TODOS need to be addressed later when `yark-pages` consideration caveats are refined
